requirements.txt
autopep8==1.4.3
certifi==2019.3.9
chardet==3.0.4
Django==2.1.5
django-filter==2.1.0
django-silk==3.0.1
djangorestframework==3.9.1
gprof2dot==2016.10.13
idna==2.8
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
Pillow==5.4.1
psycopg2==2.7.7
pycodestyle==2.5.0
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
